Split HDF5 reading tasks from API definition
