+++ src/hdf5_reader_service/tasks/__init__.py
@@ -1,0 +1,13 @@
+from .metadata import fetch_metadata
+from .search import fetch_children
+from .shapes import fetch_shapes
+from .slice import fetch_slice
+from .tree import fetch_tree
+
+__all__ = [
+    "fetch_metadata",
+    "fetch_children",
+    "fetch_shapes",
+    "fetch_slice",
+    "fetch_tree",
+]
