+++ src/hdf5_reader_service/tasks/slice.py
@@ -1,36 +1,11 @@
 import multiprocessing as mp
-import os
 from typing import Optional
 
 import h5py
-from fastapi import APIRouter
-from starlette.responses import JSONResponse
-
-from hdf5_reader_service.utils import NumpySafeJSONResponse
-
-router = APIRouter()
-
-SWMR_DEFAULT = bool(int(os.getenv("HDF5_SWMR_DEFAULT", "1")))
-
-
-# Setup blueprint route
-@router.get("/slice/")
-def get_slice(
-    path: str, subpath: str = "/", slice_info: Optional[str] = None
-) -> JSONResponse:
-    """Function that tells flask to output the metadata of the HDF5 file node.
-    The slice_info parameter should take the form
-    start:stop:steps,start:stop:steps,...
-    """
-    queue: mp.Queue = mp.Queue()
-    p = mp.Process(target=fetch_slice, args=(path, subpath, queue, slice_info))
-    p.start()
-    p.join()
-    return NumpySafeJSONResponse(queue.get())
 
 
 def fetch_slice(
-    path: str, subpath: str, slice_info: Optional[str], queue: mp.Queue
+    path: str, subpath: str, slice_info: Optional[str], swmr: bool, queue: mp.Queue
 ) -> None:
     path = "/" + path
 
@@ -45,7 +20,7 @@ def fetch_slice(
         # slices = ...
         pass
 
-    with h5py.File(path, "r", swmr=SWMR_DEFAULT, libver="latest") as f:
+    with h5py.File(path, "r", swmr=swmr, libver="latest") as f:
         if subpath in f:
             dataset = f[subpath]
             if isinstance(dataset, h5py.Dataset):
